fjern fetchmock da denne ødelegger for andre kall
fetch-mock gjør endringer på window.fetch som ødelegger funksjonalitet for andre kall i nettleseren. I dette tilfellet endret fetch-mock en post til get for et ikke fetch-mocket kall til graphql. Det er nok at fetchmock requires fra en fil så er skaden gjort.
fetch-mock bør absolutt ikke benyttes i produksjonskode, da feilen er vanskelig og tidkrevende å oppdage og løse.

fixes TAG-1952
