+++ package.json
@@ -36,7 +36,6 @@
     "crypto-browserify": "^3.12.0",
     "eslint-plugin-jsx-a11y": "^6.5.1",
     "express": "^4.17.1",
-    "fetch-mock": "^9.11.0",
     "file-saver": "^2.0.2",
     "fs-extra": "^9.1.0",
     "history": "^5.2.0",
