+++ server/mock/utils.cjs
@@ -1,0 +1,8 @@
+const delay = ms => new Promise(r => setTimeout(r, ms));
+
+const randomInt = (max) => Math.floor(Math.random() * Math.floor(max));
+
+module.exports = {
+    delay,
+    randomInt
+}
