+++ src/App/MineAnsatte/EnkeltArbeidsforhold/EnkeltArbeidsforhold.tsx
@@ -23,7 +23,8 @@ const miljø = gittMiljø<'PROD' | 'DEV' | 'LOCAL'>({
 
 const apiURL = gittMiljø({
     prod: 'https://arbeidsgiver.nav.no/arbeidsforhold/person/arbeidsforhold-api/arbeidsforholdinnslag/arbeidsgiver/{id}',
-    other: 'https://arbeidsforhold.dev.nav.no/arbeidsforhold/person/arbeidsforhold-api/arbeidsforholdinnslag/arbeidsgiver/{id}'
+    dev: 'https://arbeidsforhold.dev.nav.no/arbeidsforhold/person/arbeidsforhold-api/arbeidsforholdinnslag/arbeidsgiver/{id}',
+    other: '/arbeidsforhold/person/arbeidsforhold-api/arbeidsforholdinnslag/arbeidsgiver/{id}',
 });
 
 const EnkeltArbeidsforhold: FunctionComponent = () => {
