+++ src/index.tsx
@@ -6,31 +6,11 @@ import { init as Sentry } from '@sentry/browser';
 import 'whatwg-fetch';
 import 'abortcontroller-polyfill/dist/polyfill-patch-fetch';
 import App from './App/App';
-import { gittMiljø } from './utils/environment';
 
 Sentry({
     dsn: 'https://037b93dd39294cddb53ee3d9fad4727c@sentry.gc.nav.no/12',
     release: process.env.GIT_COMMIT_HASH || 'unknown',
     environment: window.location.hostname,
 });
-
-const kjørerLabs = gittMiljø({
-    prod: false,
-    labs: true,
-    other: false
-})
-
-if (process.env.REACT_APP_MOCK || kjørerLabs) {
-    console.log('==========================================');
-    console.log('=============== MED MOCK =================');
-    console.log('=== DETTE SKAL DU IKKE SE I PRODUKSJON ===');
-    console.log('==========================================');
-
-    require('./mocking/AaregMock');
-    require('./mocking/altinnMock');
-    require('./mocking/enkeltArbeidsforholdMock');
-    require('./mocking/BeregnetTidForArbeidsforholdMock');
-    require('./mocking/UnleashMock');
-}
 
 ReactDOM.render(<App />, document.getElementById('root'));
