--- src/mocking/Monitoring.ts
@@ -1,3 +1,0 @@
-import fetchMock from 'fetch-mock';
-
-fetchMock.get((url) => new URL(url).hostname === 'amplitude.nav.no', 200).spy();
